Add bootloader test bench
Add a new test bench that validates the boot loader can load
an application from the SD card.

The DUT is nextp8_top and the test bench includes a model of
an SD card with a SD-SPI interface.

Add read acknowledge and write acknowledge to UART status
register to avoid races.